+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Various docstrings and log messages.
 - Fix `Translator.default_fmt` typehint (is never `None`).
 - Fix inconsistent behavior of `Translator` instances that were initialized with `fetcher=TranslationMap`.
+- Fix `Translator.copy()` handling of transformers.
 
 ## [1.0.4] - 2026-01-07